Replace search overlay with navigable page
